requirements.txt

Flask==3.0.3
openai==1.35.14
python-dotenv==1.0.1
requests==2.32.3
